+++ src/units/_button.rs
@@ -6,7 +6,7 @@ use crate::config::details::{BUTTON_PINS, BUTTON_COUNT};
 pub type Button = Structure<BUTTON_COUNT>;
 
 impl Interface<BUTTON_COUNT> for Button {
-    const PINS: [u8; BUTTON_COUNT] = BUTTON_PINS;
+    const PINS: [u32; BUTTON_COUNT] = BUTTON_PINS;
     const ID: &'static str = "Buttons";
 }
